JSon Parser Library
This library is intended to use it as a Object Parser to JSON, is
capable to transform any POJO (if its fields are annotated correctly)
into a JSON Representation. Works with Primitive Types, Objects,
Collections and Maps.
There is atteched the ClienteJson Project that shows the usage if the
Library